Update headline functionality
This breaks out the headlines from the developer functionality as they are being increasingly used in editorial experimentation.

It also adds a CiF/Opinion headline option for testing
