+++ email_definitions/__init__.py
@@ -11,5 +11,6 @@ import au
 import us
 
 import developer
+import headlines
 
 import long_reads
