+++ mail_renderer.py
@@ -50,8 +50,9 @@ app = webapp2.WSGIApplication([('/daily-email/(.+)', emails.uk.DailyEmail),
                                ('/most-viewed/(.+)', emails.developer.MostViewed),
                                ('/editors-picks/(.+)', emails.developer.EditorsPicks),
                                ('/longreads/(.+)', emails.long_reads.LongReads),
-                               webapp2.Route(r'/headline', handler=emails.developer.Headline),
-                               webapp2.Route(r'/headline/film', handler=emails.developer.FilmHeadline),
-                               webapp2.Route(r'/headline/<edition>', handler=emails.developer.Headline),
+                               webapp2.Route(r'/headline', handler=emails.headlines.Headline),
+                               webapp2.Route(r'/headline/<edition:uk|us|au>', handler=emails.headlines.Headline),
+                               webapp2.Route(r'/headline/film', handler=emails.headlines.FilmHeadline),
+                               webapp2.Route(r'/headline/uk/opinion/v1', handler=emails.headlines.UKOpinion),
                                webapp2.Route(r'/', handler=handlers.Index)],
                               debug=True)
