+++ template/index.html
@@ -79,7 +79,7 @@
                 <tr>
                     <td style="padding: 10px 20px 20px 20px;">
                         <a class="story-title" href="/comment-is-free/v1" >Comment is free, v1</a> |
-                        <a class="story-title" href="/comment-is-free/v2" >v2</a> | <a class="story-title" href="/comment-is-free/v3">v3</a>
+                        <a class="story-title" href="/comment-is-free/v2" >v2</a> | <a class="story-title" href="/comment-is-free/v3">v3</a> | <a class="story-title" href="/headline/uk/opinion/v1">v1 headline</a>
                     </td>
                 </tr>
